fix fight logic
